# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "虚拟患者训练系统",
    "version": "dev"
  },
  "paths": {
...
    "/api/admin/prompts": {
      "get": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "List Prompts",
        "operationId": "list_prompts_api_admin_prompts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "purpose",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptTemplateResponse"
                  },
                  "title": "Response List Prompts Api Admin Prompts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Create Prompt",
        "operationId": "create_prompt_api_admin_prompts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/prompts/{prompt_id}": {
      "put": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Update Prompt",
        "operationId": "update_prompt_api_admin_prompts__prompt_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Delete Prompt",
        "operationId": "delete_prompt_api_admin_prompts__prompt_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/prompts/{prompt_id}/activate": {
      "post": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Activate Prompt",
        "description": "激活指定版本。prompt_id=0 表示切换到内置兜底（停用该 purpose 所有 DB 版本）。",
        "operationId": "activate_prompt_api_admin_prompts__prompt_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prompt Id"
            }
          },
          {
            "name": "purpose",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/prompts/validate": {
      "post": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Validate Prompt",
        "operationId": "validate_prompt_api_admin_prompts_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptValidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/prompts/reload": {
      "post": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Reload Prompts Endpoint",
        "operationId": "reload_prompts_endpoint_api_admin_prompts_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/prompts/sample-vars": {
      "get": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Get Sample Vars",
        "operationId": "get_sample_vars_api_admin_prompts_sample_vars_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "purpose",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleVarsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/prompts/active/preview": {
      "get": {
        "tags": [
          "Prompt管理"
        ],
        "summary": "Preview Active Prompt",
        "operationId": "preview_active_prompt_api_admin_prompts_active_preview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "purpose",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "OkResponse"
      },
...
      "PromptPreviewResponse": {
        "properties": {
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "system_prompt_raw": {
            "type": "string",
            "title": "System Prompt Raw"
          },
          "user_prompt_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt Raw"
          },
          "system_prompt_rendered": {
            "type": "string",
            "title": "System Prompt Rendered"
          },
          "user_prompt_rendered": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt Rendered"
          },
          "sample_vars": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sample Vars"
          },
          "render_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Error"
          }
        },
        "type": "object",
        "required": [
          "purpose",
          "version",
          "system_prompt_raw",
          "user_prompt_raw",
          "system_prompt_rendered",
          "user_prompt_rendered",
          "sample_vars"
        ],
        "title": "PromptPreviewResponse"
      },
      "PromptTemplateCreate": {
        "properties": {
          "purpose": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Purpose"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "system_prompt": {
            "type": "string",
            "minLength": 10,
            "title": "System Prompt"
          },
          "user_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt"
          },
          "variables": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "activate": {
            "type": "boolean",
            "title": "Activate",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "purpose",
          "system_prompt"
        ],
        "title": "PromptTemplateCreate"
      },
      "PromptTemplateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "user_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt"
          },
          "template_engine": {
            "type": "string",
            "title": "Template Engine"
          },
          "variables": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin",
            "default": false
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "version",
          "name",
          "system_prompt",
          "user_prompt",
          "template_engine",
          "variables",
          "is_active",
          "created_by",
          "remark",
          "created_at",
          "updated_at"
        ],
        "title": "PromptTemplateResponse"
      },
      "PromptTemplateUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "user_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt"
          },
          "variables": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PromptTemplateUpdate"
      },
      "PromptValidateRequest": {
        "properties": {
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "user_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt"
          },
          "variables": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "purpose",
          "system_prompt"
        ],
        "title": "PromptValidateRequest"
      },
      "PromptValidateResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "missing_vars": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Vars",
            "default": []
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "PromptValidateResponse"
      },
...
      "SampleVarsResponse": {
        "properties": {
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "vars": {
            "additionalProperties": true,
            "type": "object",
            "title": "Vars"
          }
        },
        "type": "object",
        "required": [
          "purpose",
          "vars"
        ],
        "title": "SampleVarsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}